+++ src/main/java/net/silentchaos512/gear/api/material/IMaterialCategory.java
@@ -1,0 +1,6 @@
+package net.silentchaos512.gear.api.material;
+
+@FunctionalInterface
+public interface IMaterialCategory {
+    String getName();
+}